change to int

+++ models/ncsnpp.py
@@ -243,7 +243,7 @@ class NCSNpp(nn.Module):
       # Sinusoidal positional embeddings.
       timesteps = time_cond
       # used_sigmas = self.sigmas[time_cond.long()]
-      used_sigmas = self.sigmas[time_cond]
+      used_sigmas = self.sigmas[time_cond.int()]
       temb = layers.get_timestep_embedding(timesteps, self.nf)
 
     else:
